Stop util checks isModule and isModuleProperty loading modules too aggressively.
When one of the module checks was conducted, it would try to get the checked
module immediately no matter what else was going on (though it would bail out
if the checked module *was* the module currently being loaded). Unfortunately,
this aggressive loading meant that modules could be loaded in the wrong order,
leading to dependencies being resolved incorrectly, leading to properties not
being set on the correct actual dependency objects.

In particular, this happened with the dojo/_base/declare check, which caused
dojo/_base/declare to be loaded at the time dojo/_base/kernel's first
dependency was loaded -- but dojo/_base/declare needs dojo/_base/kernel to
assign properties correctly. So it broke. But now it doesn't.